Add supplementary option to `--embed` to set PYTHONHOME.
It is possible to use the `--embed` option to build a native
binary executable.  By default this links dynamically to Python,
 but it can also be linked to a static Python library.  If you also
 copy the standard library ("/Lib" folder into the same dir as
 your new binary, you have a redistributable directory. The only
 problem is that when you run your program, Python complains
 that `prefix` and `exec_prefix` have not been set and suggests
 that you set the PYTHONHOME environment variable.   This
 commit adds a command-line option `--set-python-home`
 that will add a call to `Py_SetPythonHome("")` and the warning
 message is no longer displayed.

 One other problem with freezing in this way is that you cannot
 use the `site.py` file from CPython as-is.  It must be zeroed
 or otherwise customized because it seems that it requires the
 presence of a `build/` directory for some configuration, which
 we don't want to have to include in our redistributable directory.

The obvious alternative to adding this new option is to alter the
`cython_freeze` utility which is part of the Cython source (in the
`bin/` folder, but unfortunately not part of the source distribution
(i.e., what you get from `pip install cython`).  At the time of writing,
`cython_freeze` also doesn't set PYTHONHOME.